ath10k: fw api version 1 support
The QCA6584 firmware provided by Qualcomm uses the old
firmware format (API version 1) that does not contain
elements or magic words.

This patch adds support for these kinds of firmware files.
If the struct ath10k_hw_params element has a pointer to
a firmware file and (optionally) an otp file, these files
will be used instead of the firmware-X.bin files.

These files will be interpreted according to the old
firmware format and will not be scanned for magic words or
elements.

Signed-off-by: Erik Stromdahl <erik.stromdahl@gmail.com>