ath10k: sdio review: removed dma_buf
The dma buffer (bounce buffer) was removed.
From now on, ath10k_sdio_read_write_sync must never be called
with stack allocated buffers!

This update has introduced a few kzalloc'ed temp buffers in
order to cope with this.

At the same time, locking was fixed. The previous locking was
incomplete (locks not always held etc.).

As a consequence of this, the register spinlock was replaced
with a mutex. This was done in order to make it possible to
hold the lock while calling the sdio_ functions (that might
sleep).

Signed-off-by: Erik Stromdahl <erik.stromdahl@gmail.com>