ath10k: sdio: reduce busreq array size
The number of SDIO bus requests should never exceed the maximum number of
pending TX messages. This is because it is not possible to allocate more
MSDU id's than the TARGET_TLV_NUM_MSDU_DESC_HL limit.

Thus, the busreq array size can be set to the same value as
TARGET_TLV_NUM_MSDU_DESC_HL

This reduces the size of struct ath10k_sdio significantly.

Signed-off-by: Erik Stromdahl <erik.stromdahl@gmail.com>